+++ src/main/java/subway/auth/infrastructure/AuthenticationPrincipalConfig.java
@@ -4,7 +4,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import subway.auth.application.AuthService;
+import subway.auth.service.AuthService;
 
 import java.util.List;
