+++ src/main/java/subway/common/exception/InvalidTokenException.java
@@ -1,6 +1,6 @@
 package subway.common.exception;
 
-public class InvalidTokenException extends RuntimeException{
+public class InvalidTokenException extends RuntimeException {
 
     public static final String INVALID_TOKEN = "유효하지 않는 토큰입니다.";
