+++ src/main/java/subway/favorite/controller/FavoriteController.java
@@ -1,11 +1,11 @@
-package subway.favorite.ui;
+package subway.favorite.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import subway.auth.domain.AuthenticationPrincipal;
-import subway.favorite.application.FavoriteService;
 import subway.favorite.dto.FavoriteRequest;
 import subway.favorite.dto.FavoriteResponse;
+import subway.favorite.service.FavoriteService;
 import subway.member.domain.LoginMember;
 
 import java.net.URI;
