+++ src/main/java/subway/line/controller/LineController.java
@@ -1,11 +1,11 @@
-package subway.line.ui;
+package subway.line.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import subway.line.application.LineService;
 import subway.line.dto.LineRequest;
 import subway.line.dto.LineResponse;
 import subway.line.dto.SectionRequest;
+import subway.line.service.LineService;
 
 import java.net.URI;
 import java.util.List;
