+++ src/main/java/subway/line/dto/LineResponse.java
@@ -13,7 +13,7 @@ public class LineResponse {
     private int extraFare;
     private List<StationResponse> stations;
 
-    public LineResponse(){
+    public LineResponse() {
 
     }
