+++ src/main/java/subway/member/service/MemberService.java
@@ -1,4 +1,4 @@
-package subway.member.application;
+package subway.member.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
