+++ src/main/java/subway/station/controller/StationController.java
@@ -1,11 +1,11 @@
-package subway.station.ui;
+package subway.station.controller;
 
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import subway.station.application.StationService;
 import subway.station.dto.StationRequest;
 import subway.station.dto.StationResponse;
+import subway.station.service.StationService;
 
 import java.net.URI;
 import java.util.List;
